1. update daemon firmware for testing switch bank; 2. update Bank_WRT_Check; 3. update script of testing bank switch.

+++ Daemon/Src/mixer/mixer_processing.c
@@ -1850,7 +1850,7 @@ PT_THREAD(mixer_maintenance())
 		Gpi_Fast_Tick_Extended now = gpi_tick_fast_extended();
 
 		chirp_config.update_slot++;
-        PRINTF_CHIRP("l:%llu\n", (mx.round_deadline - now) / 16000000);
+        PRINTF_CHIRP("l:%lu\n", (uint32_t)(mx.round_deadline - now) / 16000000);
 
 		// monitor round length
 		// NOTE: we test once per slot, and STOP executes gracefully at the next slot boundary
